SLTI is a signed comparison

+++ src/simulator.c
@@ -211,7 +211,7 @@ bool execute_simulation_step(simulator* s) {
     }
     if (is_slti_instruction(&i_instruction)) {
         INFO("Instruction: SLTI %d %d %d", i_instruction.rd, i_instruction.rs1, i_instruction.imm);
-        if(read_register(s, i_instruction.rs1) < i_instruction.imm) {
+        if(((int32_t) read_register(s, i_instruction.rs1)) < i_instruction.imm) {
             write_register(s, i_instruction.rd, 1);
         }        
         else {
